bugfix in test
git-svn-id: https://svn.salilab.org/imp/trunk@11582 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/isd/test/test_MultivariateFNormalSufficient.cpp
@@ -493,7 +493,7 @@ bool test_sparseness(){
     MatrixXd Sigma(MatrixXd::Zero(100,100));
     for (int i=0; i<100; i++){
             for (int j=0; j<100; j++)
-                Sigma(i,j) = std::exp(-0.5*std::pow(std::abs(i-j),2));
+                Sigma(i,j) = std::exp(-0.5*std::pow(double(std::abs(i-j)),2));
             Sigma(i,i)+=1;
     }
